refactor: move TokenChecking
